Contracts/src/Creditworthiness.test.ts
import { CreditWorthiness } from "./Creditworthiness";

describe("CreditWorthiness Contract", () => {
  let contract: CreditWorthiness;
  
  beforeAll(async () => {
    // Deploy the CreditWorthiness contract or initialize it for testing
    contract = await deployCreditWorthinessContract();
  });

  it("should allow submitting a credit score", async () => {
    const userPublicKey = "testUserPublicKey";
    const creditScore = 700;
    await expect(contract.submitCreditScore(userPublicKey, creditScore)).resolves.not.toThrow();
  });

  it("should verify a user's creditworthiness correctly", async () => {
    const userPublicKey = "testUserPublicKey";
    const requiredScore = 650;
    const isCreditworthy = await contract.verifyCreditScore(userPublicKey, requiredScore);
    expect(isCreditworthy).toBe(true);
  });

  it("should reject a user not meeting the credit score requirement", async () => {
    const userPublicKey = "anotherTestUserPublicKey";
    const requiredScore = 750; // Higher than the submitted score
    const isCreditworthy = await contract.verifyCreditScore(userPublicKey, requiredScore);
    expect(isCreditworthy).toBe(false);
  });
});
function expect(arg0: any) {
  throw new Error("Function not implemented.");
}

function deployCreditWorthinessContract(): any {
  throw new Error("Function not implemented.");
}

function beforeAll(arg0: () => Promise<void>) {
  throw new Error("Function not implemented.");
}

